# moduls/comman.go
package moduls

// JSONResponse ...
type JSONResponse struct {
	Message string      `json:"message"`
	Data    interface{} `json:"data"`
}

// JSONErrorResponse ...
type JSONErrorResponse struct {
	Error string `json:"error"`
}